% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.cv.glmnet}
\alias{.cv.glmnet}
\title{glmnet::cv.glmnet}
\usage{
.cv.glmnet(..., nzero)
}
\arguments{
\item{...}{see \code{\link[glmnet]{cv.glmnet}}}

\item{nzero}{maximum number of non-zero coefficients\strong{:}
positive integer}
}
\value{
Object of class \code{\link[glmnet]{cv.glmnet}}.
}
\description{
Wrapper for \code{\link[glmnet]{cv.glmnet}},
with different handling of sparsity constraints.
}
\examples{
NA

}
