% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.loss}
\alias{.loss}
\title{Loss}
\usage{
.loss(y, x, family, type.measure, foldid = NULL, grouped = TRUE)
}
\arguments{
\item{y}{observed values\strong{:}
numeric vector of length \eqn{n}}

\item{x}{predicted values\strong{:}
numeric vector of length \eqn{n}}

\item{family}{character \code{"gaussian"}, \code{"binomial"}, \code{"poisson"},
\code{"mgaussian"}, or \code{"multinomial"} (to implement: \code{"cox"})}

\item{type.measure}{character \code{"deviance"}, \code{"mse"}, \code{"mae"}, \code{"class"},
or \code{"auc"}}

\item{foldid}{fold identifiers\strong{:}
integer vector of length \eqn{n},
or \code{NULL}}

\item{grouped}{logical (for \code{"cox"} only)}
}
\description{
Calculate loss from predicted and observed values
}
\examples{
NA

}
