% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.simulate}
\alias{.simulate}
\title{Simulation}
\usage{
.simulate(n, p, mode)
}
\arguments{
\item{n}{sample size\strong{:}
positive integer}

\item{p}{dimensionality\strong{:}
positive integer}

\item{mode}{character \code{"sparse"}, \code{"dense"} or \code{"blocks"}}
}
\description{
Prints object of class \link[starnet]{starnet}.
}
\examples{
NA

}
