% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{glmnet.auc}
\alias{glmnet.auc}
\title{glmnet:::auc}
\usage{
glmnet.auc(y, prob, w)
}
\arguments{
\item{y}{observed classes}

\item{prob}{predicted probabilities}

\item{w}{(ignored here)}
}
\value{
area under the ROC curve
}
\description{
Import of \code{\link[glmnet]{auc}} (internal function)
}
\examples{
NA

}
