% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.starnet}
\alias{print.starnet}
\title{Print Values}
\usage{
\method{print}{starnet}(x, ...)
}
\arguments{
\item{x}{\link[starnet]{starnet} object}

\item{...}{further arguments (not applicable)}
}
\description{
Prints object of class \link[starnet]{starnet}.
}
\examples{
set.seed(1)
n <- 30; p <- 20
y <- rnorm(n=n)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
object <- starnet(y=y,X=X)
print(object)

}
