% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate-internal}
\alias{simulate-internal}
\alias{.simulate.block}
\alias{.simulate.grid}
\alias{.simulate.mode}
\title{Simulation}
\usage{
.simulate.block(n, p, mode, family = "gaussian")

.simulate.grid(n, p, rho, pi, family = "gaussian")

.simulate.mode(n, p, mode, family = "gaussian")
}
\arguments{
\item{n}{sample size\strong{:}
positive integer}

\item{p}{dimensionality\strong{:}
positive integer}

\item{mode}{character \code{"sparse"}, \code{"dense"} or \code{"mixed"}}

\item{family}{character \code{"gaussian"}, \code{"binomial"} or \code{"poisson"}}

\item{rho}{correlation\strong{:}
numeric between \eqn{0} and \eqn{1}}

\item{pi}{effects\strong{:}
numeric between \eqn{0} (sparse) and \eqn{1} (dense)}
}
\value{
List of vector \code{y} and matrix \code{X}.
}
\description{
Functions for simulating data
}
\examples{
NA
}
