% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cv.starnet}
\alias{cv.starnet}
\title{Model comparison}
\usage{
cv.starnet(y, X, family = "gaussian", nalpha = 21, alpha = NULL,
  nfolds.ext = 10, nfolds.int = 10, foldid.ext = NULL,
  foldid.int = NULL, type.measure = "deviance", alpha.meta = 1,
  grouped = TRUE, nzero = NULL, intercept = NULL,
  upper.limit = NULL, unit.sum = NULL, ...)
}
\arguments{
\item{y}{numeric response\strong{:}
vector of length \eqn{n}}

\item{X}{covariates\strong{:}
numeric matrix with \eqn{n} rows (samples)
and \eqn{p} columns (variables)}

\item{family}{character "gaussian", "binomial" or "poisson"}

\item{nalpha}{number of \code{alpha} values}

\item{alpha}{elastic net mixing parameters\strong{:}
vector of values between \eqn{0} (ridge) and \eqn{1} (lasso);
or \code{NULL} (equidistance)}

\item{nfolds.ext, nfolds.int, foldid.ext, foldid.int}{(number of) external/internal folds}

\item{type.measure}{loss function\strong{:}
character "deviance", "class", "mse" or "mae"
(see \code{\link[glmnet]{cv.glmnet}})}

\item{alpha.meta}{elastic net mixing parameters for stacking\strong{:}
vector of values between \eqn{0} (ridge) and \eqn{1} (lasso),
\emph{see details}}

\item{grouped}{logical}

\item{nzero}{number of non-zero coefficients\strong{:}
scalar/vector including positive integer(s) or \code{NA};
or \code{NULL} (no posthoc feature selection)}

\item{intercept}{settings for meta-learner\strong{:} logical}

\item{upper.limit}{settings for meta-learner\strong{:} logical}

\item{unit.sum}{settings for meta-learner\strong{:} logical}

\item{...}{further arguments (not applicable)}
}
\description{
Compares stacked elastic net, tuned elastic net, ridge and lasso.
}
\examples{
set.seed(1)
n <- 50; p <- 20
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
y <- rnorm(n=n,mean=rowSums(X[,1:20]))
loss <- cv.starnet(y=y,X=X)

}
