% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.starnet}
\alias{print.starnet}
\title{Print Values}
\usage{
\method{print}{starnet}(x, ...)
}
\arguments{
\item{x}{\link[starnet]{starnet} object}

\item{...}{further arguments (not applicable)}
}
\value{
Prints "stacked gaussian/binomial/poisson elastic net".
}
\description{
Prints object of class \link[starnet]{starnet}.
}
\examples{
\dontshow{
if(!grepl('SunOS',Sys.info()['sysname'])){
set.seed(1)
n <- 50; p <- 100
y <- rnorm(n=n)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
object <- starnet(y=y,X=X)
print(object)}}
\donttest{
set.seed(1)
n <- 50; p <- 100
y <- rnorm(n=n)
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
object <- starnet(y=y,X=X)
print(object)}

}
