% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{starnet-package}
\alias{starnet-package}
\title{Stacked Elastic Net Regression}
\description{
The R package \code{starnet} implements stacked elastic net regression.
The elastic net generalises ridge and lasso regularisation.
Instead of fixing or tuning the mixing parameter alpha,
we combine multiple alphas by stacked generalisation.
}
\details{
Use function \code{\link{starnet}} for model fitting.
Type \code{library(starnet)} and then \code{?starnet} or
\code{help("starnet)"} to open its help file.

See the vignette for further examples.
Type \code{vignette("starnet")} or \code{browseVignettes("starnet")}
to open the vignette.
}
\references{
A Rauschenberger, E Glaab, and MA van de Wiel (2020).
"Predictive and interpretable models via the stacked elastic net".
\emph{Bioinformatics}. In press.
\doi{10.1093/bioinformatics/btaa535}.
\email{armin.rauschenberger@uni.lu}
}
\keyword{documentation}
